Modified   E-LinkStatusConfig.xml Modified   msvc/E-LinkStatus.vcxproj Modified   src/StatusManager.cpp Modified   src/StatusManager.hpp Modified   src/StatusPing.cpp Modified   src/StatusPing.hpp Modified   src/fonts/KS0108.cpp Modified   src/fonts/KS0108.hpp Modified   src/painter/Painter.cpp Modified   src/painter/Painter.hpp Modified   src/painter/PainterFont.hpp Add        .gitignore

+++ src/fonts/KS0108.cpp
@@ -8,13 +8,13 @@ using namespace std;
 
 FontPainterKS0108::FontPainterKS0108( FontData fontData, DrawPixel drawPixel ) : FontPainter( fontData, drawPixel ) {}
 
-FontPainter::Size FontPainterKS0108::drawChar( unsigned startX, unsigned startY, char character, bool color )
+FontPainter::Size FontPainterKS0108::drawChar( size_t startX, size_t startY, char character, bool color )
 {
 	return drawChar( startX, startY, character, color, m_drawPixel );
 }
 
 FontPainter::Size FontPainterKS0108::drawChar(
-	unsigned startX, unsigned startY, char character, bool color, DrawPixel drawPixel )
+	size_t startX, size_t startY, char character, bool color, DrawPixel drawPixel )
 {
 	const auto chacterIndex = static_cast< uint8_t >( character ) - static_cast< uint8_t >( m_fontData.startCharacter );
 
@@ -35,13 +35,13 @@ FontPainter::Size FontPainterKS0108::drawChar(
 	return { charData->width, m_fontData.height };
 }
 
-FontPainter::Size FontPainterKS0108::drawString( unsigned startX, unsigned startY, std::string text, bool color )
+FontPainter::Size FontPainterKS0108::drawString( size_t startX, size_t startY, std::string text, bool color )
 {
 	return drawString( startX, startY, text, color, m_drawPixel );
 }
 
 FontPainter::Size FontPainterKS0108::drawString(
-	unsigned startX, unsigned startY, std::string text, bool color, DrawPixel drawPixel )
+	size_t startX, size_t startY, std::string text, bool color, DrawPixel drawPixel )
 {
 	FontPainter::Size totalSize{ 0, 0 };
 	size_t nextCharXPosition{ 0 };
@@ -65,6 +65,5 @@ FontPainter::Size FontPainterKS0108::getStringSize( std::string text )
 {
 	return drawString( 0, 0, text, false, DrawPixel() );
 }
-
 
 // end: --------------------- FontPainterKS0108 -------------------
