+++ src/painter/PainterFont.hpp
@@ -5,13 +5,13 @@
 class FontPainter
 {
 protected:
-	using DrawPixel = std::function< void( unsigned, unsigned, bool ) >;
+	using DrawPixel = std::function< void( size_t, size_t, bool ) >;
 
 public:
 	struct Size
 	{
-		unsigned width;
-		unsigned height;
+		size_t width;
+		size_t height;
 	};
 
 protected:
@@ -19,10 +19,10 @@ protected:
 	virtual ~FontPainter() = default;
 
 public:
-	virtual Size drawChar( unsigned startX, unsigned startY, char character, bool color )	 = 0;
-	virtual Size drawString( unsigned startX, unsigned startY, std::string text, bool color ) = 0;
-	virtual Size getCharSize( char character )												  = 0;
-	virtual Size getStringSize( std::string text )											  = 0;
+	virtual Size drawChar( size_t startX, size_t startY, char character, bool color )	 = 0;
+	virtual Size drawString( size_t startX, size_t startY, std::string text, bool color ) = 0;
+	virtual Size getCharSize( char character )											  = 0;
+	virtual Size getStringSize( std::string text )										  = 0;
 
 protected:
 	DrawPixel m_drawPixel;
